릴리즈 타겟으로 커밋

+++ Server/GameWebServer/Modules/Serializer/PacketDeserializer.cs
@@ -39,7 +39,7 @@ namespace GameWebServer.Modules.Serializer
 
             var body = Encoding.UTF8.GetString(bytes, ProtocolSize, bytes.Length - ProtocolSize);
 
-            _interServerFuncHandler.Process((int)protocol, body);
+            _interServerFuncHandler.Process(protocol, body);
         }
     }
 }
